requirements.txt
scikit-learn
numpy
pandas
matplotlib
jupyter
feedparser==6.0.11
python-dateutil==2.9.0.post0
tqdm==4.66.4
PyYAML==6.0.2
google-api-python-client==2.137.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
python-docx==1.1.2
openai==1.37.1
requests>=2.32.0
python-dotenv>=1.0.1